+++ file_options.go
@@ -12,11 +12,12 @@ import (
 
 // Options define the behavior of `WriteReaderAtomic()`, etc.
 type Options struct {
-	defaultFileMode os.FileMode // default 0
-	fileMode        os.FileMode // default 0
-	flags           int         // default 0
-	keepFileMode    bool        // default false
-	setSymlinkOwner bool        // default false
+	defaultFileMode os.FileMode  // default 0
+	fileMode        os.FileMode  // default 0
+	flags           int          // default 0
+	keepFileMode    bool         // default false
+	readOnlyMode    ReadOnlyMode // default 0
+	setSymlinkOwner bool         // default false
 }
 
 // Option functions modify Options.
@@ -50,6 +51,20 @@ func WithFlags(flags int) Option {
 func WithKeepFileMode(keep bool) Option {
 	return func(opts *Options) {
 		opts.keepFileMode = keep
+	}
+}
+
+// WithReadOnlyMode is used to determine if/when to set a file's read-only
+// (RO) attribute on Windows. The following values are supported:
+// ReadOnlyModeIgnore do not set a file's RO attribute, and ignore if it's set.
+// ReadOnlyMaskSet    set a file's RO attribute if the file's FileMode has the
+//
+//	user writable bit set.
+//
+// ReadOnlyMaskReset  do not set a file's RO attribute, and if it's set, reset it.
+func WithReadOnlyMode(mode ReadOnlyMode) Option {
+	return func(opts *Options) {
+		opts.readOnlyMode = mode
 	}
 }